add profiles

+++ reegis_hp/de21/ew.py
@@ -34,37 +34,57 @@ STATES = {
 
 
 def get_ew_shp_file(year):
-    # url = ('http://www.geodatenzentrum.de/auftrag1/archiv/vektor/' +
-    #        'vg250_ebenen/{0}/vg250-ew_{0}-12-31.geo84.shape'.format(year) +
-    #        '.{0}.zip')
-    url = cfg.get('download', 'url_geodata_ew').format(year=year,
-                                                       var1='{0}')
-    filename_zip = os.path.join(cfg.get('paths', 'general'),
-                                cfg.get('general_sources', 'vg250_ew_zip'))
-    msg = t.download_file(filename_zip, url.format('ebene'))
-    if msg == 404:
-        logging.warning("Wrong URL. Try again with different URL.")
-        t.download_file(filename_zip, url.format('ebenen'), overwrite=True)
-    zip_ref = zipfile.ZipFile(filename_zip, 'r')
-    zip_ref.extractall(cfg.get('paths', 'general'))
-    zip_ref.close()
-    subs = next(os.walk(cfg.get('paths', 'general')))[1]
-    mysub = None
-    for sub in subs:
-        if 'vg250' in sub:
-            mysub = sub
-    pattern_path = os.path.join(cfg.get('paths', 'general'),
-                                mysub,
-                                'vg250-ew_ebenen',
-                                'VG250_VWG*')
-    for file in glob.glob(pattern_path):
-        file_new = os.path.join(cfg.get('paths', 'general'),
-                                'VG250_VWG_' + str(year) + file[-4:])
-        shutil.copyfile(file, file_new)
+    if year < 2009:
+        logging.error("Shapefile with inhabitants are available since 2009.")
+        logging.error("Try to find another source to get older data sets.")
+        raise AttributeError('Years < 2009 are not allowed in this function.')
 
-    shutil.rmtree(os.path.join(cfg.get('paths', 'general'), mysub))
+    outshp = os.path.join(cfg.get('paths', 'general'),
+                          'VG250_VWG_' + str(year) + '.shp')
 
-    os.remove(filename_zip)
+    if not os.path.isfile(outshp):
+        url = cfg.get('download', 'url_geodata_ew').format(year=year,
+                                                           var1='{0}')
+        filename_zip = os.path.join(cfg.get('paths', 'general'),
+                                    cfg.get('general_sources', 'vg250_ew_zip'))
+        msg = t.download_file(filename_zip, url.format('ebene'))
+        if msg == 404:
+            logging.warning("Wrong URL. Try again with different URL.")
+            t.download_file(filename_zip, url.format('ebenen'), overwrite=True)
+
+        zip_ref = zipfile.ZipFile(filename_zip, 'r')
+        zip_ref.extractall(cfg.get('paths', 'general'))
+        zip_ref.close()
+        subs = next(os.walk(cfg.get('paths', 'general')))[1]
+        mysub = None
+        for sub in subs:
+            if 'vg250' in sub:
+                mysub = sub
+        pattern_path = list()
+
+        pattern_path.append(os.path.join(cfg.get('paths', 'general'),
+                                         mysub,
+                                         'vg250-ew_ebenen',
+                                         'VG250_VWG*'))
+        pattern_path.append(os.path.join(cfg.get('paths', 'general'),
+                                         mysub,
+                                         'vg250-ew_ebenen',
+                                         'vg250_vwg*'))
+        pattern_path.append(os.path.join(cfg.get('paths', 'general'),
+                                         mysub,
+                                         'vg250_ebenen-historisch',
+                                         'de{0}12'.format(str(year)[-2:]),
+                                         'vg250_vwg*'))
+
+        for pa_path in pattern_path:
+            for file in glob.glob(pa_path):
+                file_new = os.path.join(cfg.get('paths', 'general'),
+                                        'VG250_VWG_' + str(year) + file[-4:])
+                shutil.copyfile(file, file_new)
+
+        shutil.rmtree(os.path.join(cfg.get('paths', 'general'), mysub))
+
+        os.remove(filename_zip)
 
 
 def get_ew_by_region(spatial, year):
@@ -134,4 +154,4 @@ if __name__ == "__main__":
     spatial_file_fs = os.path.join(cfg.get('paths', 'geometry'),
                                    cfg.get('geometry', 'federalstates_polygon'))
     spatial_dfs = pd.read_csv(spatial_file_fs, index_col='gen')
-    print(get_ew_by_region(spatial_dfs, 2014))
+    print(get_ew_by_region(spatial_dfs, 2009))
